back/scripts/producer/requirements.txt
confluent-kafka==2.3.0
requests==2.31.0
python-dotenv
pandas
pycoingecko
python-binance
